ml_backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
xgboost>=2.1.0
scikit-learn>=1.5.0
pandas>=2.2.0
numpy>=1.26.0
requests>=2.32.0
joblib>=1.4.0
python-dotenv>=1.0.0
apscheduler>=3.10.0
supabase>=2.0.0
